1) Out put file ready:
- printing matrix of paths representing parameter of paper called
ro(p,e) that represents for each path p if it has edge e or not.

- printing matrix of transmissions representing parameter of paper
called ro(r,e) that represents for each transmission r if it has edge e
or not.

All of this assumed that in input file we specify the graphy with all
its edges and for each transmission the path that follows, each path is
just a list of nodes separated by white spaces.

2) Class Edge ready with parameter collection of transmissions, getter
and setter and method to get all the free slices of that edge.